requirements.txt
streamlit
langchain
langchain-google-genai
langchain-community
faiss-cpu
python-dotenv
tavily-python
nest_asyncio